Add notes related to phobos.sys.meta and std.meta. (#36)
It seemed like it would be a good idea to document which symbols in
Phobos v2 map to which ones in Phobos v3, as well as to document which
symbols are _not_ being ported to Phobos v3 (and why). In addition, it
seems sensible to document at least some of the design decisions. And it
will also help me keep track of where things stand for myself when
figuring out what I still need to do - though who knows whether anyone
else will be doing this kind of documentation later for the stuff that
they work on.

Either way, when Phobos v3 gets released, we _are_ going to need to put
together information for users somewhere for what ended up where in
Phobos v3 so that they can port their Phobos v2 code to v3 if they so
desire (and generally find stuff in v3 that they're familiar with in
v2). And documentation like this commit has can at least provide the
basis for that even if it's not what we're going to give to users
(presumably, that will need to be in ddoc on dlang.org).

So, I've started a section of the repo for providing notes for each
module as we add them (as well as notes for the old modules so that we
can provide links to the new symbols and provide information about
symbols that are not getting ported over). So, the intention at least is
that each public module in Phobos v2 and each public module in Phobos v3
will eventually end up with a corresponding file under module_notes, but
we'll see.